performetric sync

+++ database_schemas/performetric_database_schema.sql
@@ -1,7 +1,7 @@
 # to be executed in the aware study database
 
 CREATE TABLE IF NOT EXISTS performetric_fatigue_report(
-  _id INT NOT NULL PRIMARY KEY AUTO_INCREMENT,
+  _id BIGINT NOT NULL PRIMARY KEY AUTO_INCREMENT,
   user VARCHAR(255) NOT NULL,
   fatigue_avg FLOAT,
   minutes_no_fatigue INT,
@@ -10,5 +10,5 @@ CREATE TABLE IF NOT EXISTS performetric_fatigue_report(
   rest_breaks INT,
   fatigue_messages INT,
  `from` DATETIME NOT NULL,
-  `to` DATETIME NOT NULL
+ `to` DATETIME NOT NULL
 );
